+++ app/home/html/home.html
@@ -1,1 +1,1 @@
-<h1 ng-bind="$page.systemTitle"></h1>
+<h2 ng-bind="$page.title"></h2>
