+++ app/main/controller/main.controller.js
@@ -1,0 +1,10 @@
+(function(){
+    "use strict";
+
+    angular.module("app.common")
+        .controller("mainController", mainController);
+
+    function mainController(){
+
+    }
+})();
