+++ app/main/html/main.html
@@ -4,11 +4,13 @@
         <title ng-bind="$page.systemTitle"></title>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <!--link href="css/style.css" rel="stylesheet" /-->
+        <link href="css/libs.min.css" rel="stylesheet" />
+        <link href="css/styles.min.css" rel="stylesheet" />
         <script src="js/libs.min.js" type="text/javascript"></script>
         <script src="js/scripts.min.js" type="text/javascript"></script>
     </head>
     <body>
+        <h1 ng-bind="$page.systemTitle"></h1>
         <div ui-view></div>
     </body>
 </html>
