+++ app/scss/styles.scss
@@ -1,0 +1,5 @@
+body
+{
+    font-family: Arial, Tahoma, Helvetica, sans-serif;
+    font-size: 16px;
+}
