+++ package.json
@@ -10,20 +10,15 @@
     "gulp": "^3.9.1",
     "gulp-autoprefixer": "^3.1.1",
     "gulp-clean": "^0.3.2",
-    "gulp-clean-css": "^2.0.12",
     "gulp-concat": "^2.6.0",
-    "gulp-for-compass": "0.0.2",
-    "gulp-htmlmin": "^2.0.0",
     "gulp-jshint": "^2.0.1",
     "gulp-rename": "^1.2.2",
     "gulp-sass": "^2.3.2",
     "gulp-sourcemaps": "^2.0.0",
     "gulp-uglify": "^2.0.0",
     "gulp-util": "^3.0.7",
-    "gulp-watch": "^4.3.10",
     "jshint": "^2.9.3",
     "run-sequence": "^1.2.2",
-    "serve-static": "^1.11.1",
-    "yargs": "^6.0.0"
+    "yargs": "^6.1.1"
   }
 }
