+++ server.js
@@ -1,12 +1,18 @@
-var express = require('express');
+"use strict";
+
+var express = require("express");
 var http = require("https");
+var port = 3000;
+var gulp = require("gulp");
+require("./gulpfile.js");
+
 var app = express();
 
-app.use(express.static(__dirname));
+app.use(express.static('dist'));
 
 var pages = {
-    login: __dirname + '/security/html/login.html',
-    app: __dirname + '/app/base/html/main.html'
+    login: __dirname + "/security/html/login.html",
+    app: __dirname + "/dist/main/html/main.html"
 }
 
 function getOptions(url, path)
@@ -22,7 +28,7 @@ function getOptions(url, path)
     };
 }
 
-app.all("/cartola/mercado/status", function(req, res, next){
+app.get("/cartola/mercado/status", function(req, res, next){
     var options = getOptions("/mercado/status");
     var rq = http.request(options, function (r) {
         var chunks = [];
@@ -31,14 +37,13 @@ app.all("/cartola/mercado/status", function(req, res, next){
         });
         r.on("end", function () {
             var body = Buffer.concat(chunks);
-            //console.log(body.toString());
             res.send(body.toString());
         });
     });
     rq.end();
 });
 
-app.all("/cartola/pos-rodada/destaques", function(req, res, next){
+app.get("/cartola/pos-rodada/destaques", function(req, res, next){
     var options = getOptions("/pos-rodada/destaques");
     var rq = http.request(options, function (r) {
         var chunks = [];
@@ -47,17 +52,22 @@ app.all("/cartola/pos-rodada/destaques", function(req, res, next){
         });
         r.on("end", function () {
             var body = Buffer.concat(chunks);
-            //console.log(body.toString());
             res.send(body.toString());
         });
     });
     rq.end();
 });
 
-//app.all("/")
-
-var port = 3000;
+app.get("/", function(req, res, next){
+    res.sendFile(__dirname + "/dist/main/html/main.html");
+});
 
 app.listen(port);
 
-console.log('Server running on ' + port + '...');
+console.log("Servidor executando na porta " + port + "...");
+
+/* Gulp */
+
+if (gulp.hasTask("build:dev")) { 
+    gulp.start("build:dev");
+}
